Fix redirect on dashboard page after reload when logged in
This was due to not waiting for authentication from firebase before
redirecting to the signup page. Now, the value `undefined` is used when
no authentication information is available and the special string
'NOT_LOGGED_IN' is used to denote the state where it has been determined
that the user is not logged in.