Fix zero geocoding results to resolve to empty array
Before applying this fix, the status 'ZERO_RESULTS' would mean
that we return a rejected Promise, which will be caught in LocationInput
and reported as an error to Sentry.

With this fix we resolve the Promise with an empty array and default to
rejecting the Promise for all other statuses.

Close #189