Catch errors when parsing entries
Some entries have lat/lon zero which leads to an error
when we try to extract the data with the firebase library.

Instead of crashing when one entry has this problem, we
are now filtering out those invalid entries and logging
an error instead.

Note that in production we should not have those entries
and they indicate that something is wrong in another place,
which is why I've added reporting to sentry if any of those
invalid entries should occur.

Close #220